pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-cs"
version = "0.1.0"
description = "MCP configuration based server"
authors = [
    {name = "Yiheng Tao", email = "example@example.com"}
]
requires-python = ">=3.11"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Core dependencies
    "mcp>=1.7.1",  # Model Context Protocol Python SDK
    
    # Vector database
    "chromadb>=0.4.22",  # Vector database for semantic search
    
    # YAML parser
    "pyyaml>=6.0.1",  # YAML parser
    
    # Process utilities
    "psutil>=5.9.5",  # Process utilities for monitoring and management
    
    # HTTP/SSE server
    "uvicorn>=0.25.0",  # ASGI server
    "starlette>=0.37.0",  # ASGI framework
    "httpx>=0.26.0",  # HTTP client
    "httpx-sse>=0.4.0",  # SSE client for httpx
    
    # Webdriver tools
    "selenium==4.18.1",
    "webdriver-manager==4.0.1",
    "playwright>=1.42.0",  # Modern browser automation
    
    # Text processing
    "trafilatura>=1.6.2",  # Web content extraction
    "beautifulsoup4>=4.12.0",  # HTML parsing for extraction
    "markdownify>=0.11.6",  # HTML to markdown conversion
    
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    
    # Azure Data Explorer (Kusto)
    "azure-kusto-data>=5.0.0",  # Azure Data Explorer client
    "azure-identity>=1.15.0",  # Azure authentication
    
    # Embedding models
    "sentence-transformers>=2.7.0",  # Lightweight embedding model for tests
    "detect-secrets>=1.4.0",  # Secret scanning
    
    # Image processing (for chart extractor)
    "easyocr>=1.7.1",  # Fast OCR for text extraction from images
    "opencv-python>=4.8.0",  # OpenCV for image processing
    "numpy>=1.24.0",  # NumPy for numerical operations
    "pillow>=9.5.0",  # PIL for image handling
    
    # Log compression and parsing
    "msgpack>=1.0.5",  # MessagePack serialization
    "zstandard>=0.21.0",  # Zstandard compression
    "logparser==0.8.4",  # Log parsing library
    "click>=8.1.3",  # Command line interface
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
]

[tool.setuptools]
packages = ["server", "mcp_tools", "mcp_core", "config", "utils"]

# This tells uv/pip to install the local packages in development mode
[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function" 